+++ cli/ballerina-cli/src/main/java/io/ballerina/cli/task/RestoreCachedArtifactsTask.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright (c) 2025, WSO2 LLC. (https://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.ballerina.cli.task;
+
+import io.ballerina.cli.cmd.CommandUtil;
+import io.ballerina.projects.Project;
+import io.ballerina.projects.util.ProjectConstants;
+
+import java.io.IOException;
+
+import static io.ballerina.cli.launcher.LauncherUtils.createLauncherException;
+
+public class RestoreCachedArtifactsTask implements Task {
+    @Override
+    public void execute(Project project) {
+        try {
+            CommandUtil.copyOneDirectoryUp(project.targetDir().resolve(ProjectConstants.EXEC_BACKUP_DIR_NAME));
+        } catch (IOException e) {
+            throw createLauncherException("unable to restore the cache: " + e.getMessage());
+        }
+    }
+}
